backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
pydantic==2.11.0
pydantic-settings==2.4.0
python-multipart==0.0.9
